apps/voice-agent/requirements.txt
livekit-agents==1.4.6
livekit-plugins-openai==1.4.6
livekit-plugins-silero==1.4.6
httpx==0.28.1
python-dotenv==1.0.1
